- extended codechecker script (with --stats option) + plotter

+++ codechecker.py
@@ -3,14 +3,17 @@ import glob
 import sys
 import os
 
-def codechecker(filename, tmpdir, verbose=True):
+def codechecker(filename, tmpdir, pullcode=True, verbose=True):
 	df = pd.read_csv(filename, names=["id", "fqid", "vendor", "deployments", "tags", "description", "url"])
 
 	urllist = list(df["url"])
 	hascloned = {}
 	success = 0
 	failure = 0
+	unpulled = 0
 	dupe = 0
+	otherurls = {}
+	nourls = 0
 	for i, url in enumerate(urllist):
 		if "github.com" in str(url):
 			urlstem = "/".join(url.split("/")[:5])
@@ -18,7 +21,7 @@ def codechecker(filename, tmpdir, verbose=True):
 			if urlstem in hascloned:
 				print("reuse {} ({})".format(urlstem, hascloned[urlstem]))
 				dupe += 1
-			else:
+			elif pullcode:
 				if os.path.isdir("{}/_codechecker/{}".format(tmpdir, i)):
 					print("clone-update {}...".format(urlstem))
 					origdir = os.getcwd()
@@ -38,29 +41,59 @@ def codechecker(filename, tmpdir, verbose=True):
 					else:
 						success += 1
 						hascloned[urlstem] = i
-			os.makedirs("{}/_codefolders".format(tmpdir), exist_ok=True)
-			origdir = None
-			if urlpath.startswith("tree"):
-				tree, treename, *rest = urlpath.split("/")
-				urlpath = "/".join(rest)
-				origdir = os.getcwd()
-				os.chdir("{}/_codechecker/{}".format(tmpdir, hascloned[urlstem]))
-				os.system("git checkout {}".format(treename))
-				os.chdir(origdir)
-			os.system("rm -rf {}/_codefolders/{}".format(tmpdir, i))
-			os.system("cp -r {}/_codechecker/{}/{} {}/_codefolders/{}".format(tmpdir, hascloned[urlstem], urlpath, tmpdir, i))
+			else:
+				unpulled += 1
+				hascloned[urlstem] = i
+
+			if pullcode:
+				os.makedirs("{}/_codefolders".format(tmpdir), exist_ok=True)
+				origdir = None
+				if urlpath.startswith("tree"):
+					tree, treename, *rest = urlpath.split("/")
+					urlpath = "/".join(rest)
+					origdir = os.getcwd()
+					os.chdir("{}/_codechecker/{}".format(tmpdir, hascloned[urlstem]))
+					os.system("git checkout {}".format(treename))
+					os.chdir(origdir)
+				os.system("rm -rf {}/_codefolders/{}".format(tmpdir, i))
+				os.system("cp -r {}/_codechecker/{}/{} {}/_codefolders/{}".format(tmpdir, hascloned[urlstem], urlpath, tmpdir, i))
 			#if origdir:
 			#	os.chdir("_codechecker/{}".format(hascloned[urlstem]))
 			#	os.system("git checkout master")
 			#	os.chdir(origdir)
+		else:
+			if pd.isnull(url):
+				nourls += 1
+			else:
+				urlstem = url
+				if "/" in url:
+					urlstem = url.split("/")[2]
+				otherurls[urlstem] = otherurls.get(urlstem, 0) + 1
 
-	print("failures {} + success {} = unique repos {} + dupes {} = github {} + other {}".format(failure, success, failure + success, dupe, failure + success + dupe, len(urllist) - failure - success - dupe))
+	print("failures {} + success {} = unique github repos {} + dupes {} = github {} + other {} + none {} = total {}".format(failure, success, failure + success + unpulled, dupe, failure + success + unpulled + dupe, len(urllist) - failure - success - unpulled - dupe - nourls, nourls, len(urllist)))
+	print("others", otherurls)
+
+	return len(urllist), failure, success, unpulled, dupe, nourls
 
 tmpdir = "."
+stats = False
 if len(sys.argv) == 2:
 	tmpdir = sys.argv[1]
+	if sys.argv[1] == "--stats":
+		stats = True
 
 #filenames = glob.glob("autostats/autocontents-*.csv")
 filenames = glob.glob("autocontents-*.csv")
 filenames.sort()
-codechecker(filenames[-1], tmpdir)
+
+if not stats:
+	codechecker(filenames[-1], tmpdir)
+else:
+	f = open("codechecker.csv", "w")
+	print("#date,total,github-unique,github-dupe,other,none", file=f)
+	for filename in filenames:
+		date = filename.replace("autocontents-", "").replace(".csv", "")
+		total, fign, sign, unpulled, dupe, nourls = codechecker(filename, None, False)
+		other = total - unpulled - dupe - nourls
+		print("{},{},{},{},{},{}".format(date, total, unpulled, dupe, other, nourls), file=f)
+	f.close()
